bug fixes; add foreign key and change datatype. added: migration and seeder for departemen master, fakultas master, and pekerjaan master

+++ database/migrations/2020_04_26_064427_create_info_mahasiswas_table.php
@@ -21,7 +21,7 @@ class CreateInfoMahasiswasTable extends Migration
             $table->tinyInteger('fakultas');
             $table->tinyInteger('departemen');
             $table->integer('tahun_masuk');
-            $table->integer('ipk');
+            $table->decimal('ipk', 3, 2);
             $table->string('transkrip');
             $table->integer('spp');
             $table->timestamps();
